+++ apps/studio/src/app/graph/graph-hooks.ts
@@ -15,7 +15,6 @@ export const usePackagesQuery = () => {
   );
 
   useEffect(() => {
-    console.log({ isLoading });
     setLoading(isLoading);
   }, [isLoading, setLoading]);
