+++ packages/ui-graph/commonality.json
@@ -1,1 +1,1 @@
-{ "tags": ["ui"] }
+{"tags":[]}
